refactor: ErrorCard & StatusCard
