+++ src/components/index.js
@@ -1,8 +1,8 @@
 export { default as CardInput } from './cards/CardInput';
 export { default as CardValue } from './cards/CardValue';
-export { default as CardParams } from './cards/CardParams';
-export { default as CardSteps } from './cards/CardSteps';
 export { default as CardStatus } from './cards/CardStatus';
+export { default as CardErrors } from './cards/CardErrors';
+export { default as CardSteps } from './cards/CardSteps';
 
 export { default as DividerCustom } from './common/DividerCustom';
 export { default as NavBar } from './common/NavBar';
